Separate classes for activities; added bin,obj,.suo to gitignore; first pass at unit tests
