+++ HelloInput/SimpleSum.cs
@@ -1,0 +1,28 @@
+﻿using System.Activities;
+using System.ComponentModel;
+
+namespace HelloInput
+{
+    public class SimpleSum : CodeActivity
+    {
+        [Category("Input")]
+        [RequiredArgument]
+        public InArgument<double> FirstNumber { get; set; }
+
+        [Category("Input")]
+        public InArgument<double> SecondNumber { get; set; }
+
+        [Category("Output")]
+        public OutArgument<double> ResultNumber { get; set; }
+
+        protected override void Execute(CodeActivityContext context)
+        {
+            var firstNumber = FirstNumber.Get(context);
+            var secondNumber = SecondNumber.Get(context);
+            var result = firstNumber + secondNumber;
+            ResultNumber.Set(context, result);
+        }
+
+
+    }
+}
